+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 - Added `RxCollection().bulkUpsert()`
 - Added `RxStorageInstance.cleanup()`
 - Added cleanup plugin
+- dev-mode: Add check to ensure all top-level fields in a query are defined in the schema.
 
 ### 11.6.0 (4 February 2022)